refactor(st-09061): split tool type contracts
