+++ packages/core/src/tools/types-category.ts
@@ -1,0 +1,13 @@
+/**
+ * Categories for organizing tools.
+ */
+export enum ToolCategory {
+  FILE_SYSTEM = 'file-system',
+  WEB = 'web',
+  CODE = 'code',
+  DATABASE = 'database',
+  API = 'api',
+  UTILITY = 'utility',
+  CUSTOM = 'custom',
+  SKILLS = 'skills',
+}
