+++ planning/epics-and-stories.md
@@ -1888,7 +1888,7 @@
 **Priority:** P2 (Medium)
 **Estimate:** 4 hours
 **Dependencies:** ST-09050
-**Status:** Ready
+**Status:** In Progress
 
 **Acceptance criteria:**
 - [ ] `packages/core/src/tools/types.ts` is reduced below the 300 line planning cutoff by extracting focused internal modules for tool result types, execution/input contracts, schema-facing shared helpers, and public utility types behind a stable facade, and the extracted production modules must also stay below `300` lines unless the story documents and justifies an explicit exception.
