+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-06-08
-**Active Story:** ST-09061 - Modularize Core Tool Types and Tests (Ready)
+**Last Updated:** 2026-06-10
+**Active Story:** ST-09061 - Modularize Core Tool Types and Tests (In Progress)
 
 ---
